backend/requirements.txt
Django==4.2.15
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.1.0
whitenoise==6.6.0
django-cockroachdb==4.2
dj-database-url==2.1.0
channels==4.0.0
channels-redis==4.1.0
gunicorn==21.2.0
requests==2.31.0
django-storages==1.14.2
boto3==1.34.0